Add better support for WikiLinks and other goodies (#2)
This PR provides better support for WikiLinks as a hook than the existing plugins for Mkdocs. One major issue solved here is the snippet extraction before Wikilink resolution. The new hook also reports Wikilinks that point to no existing pages, along with other goodies.

As with the other chores in this PR, it organizes the other hooks better, separates the diff routines from the Juvix preprocessing, and makes other improvements.