Add new hook for image (pre)processing (#7)
Images are all stored in `docs/images`.
Image insertion would resolve the file path for any image. This means,
you only need to refer to its filename when including an image in your
file. For example,

```
![caption...](filename.png)
```

The new hook also processes the `.dot` file. These files must also be in
`docs/images`. If you want to include a `.dot` file, do as follows:


```
![caption...](mydotfile.dot.svg)
```

Notice the `.svg` at the end.

As a result of this new hook, we can detect references to non-exiting
image files and errors in converting dot files to SVG files. It also
simplifies the way images are included in the text.