Refactor file and folder names: add snake_case convention (#115)
This PR fixes the following issues:

- rename all files/folders to conform to the snake_case convention. This
is the new standard. To run this refactor, I wrote [this
script](https://gist.github.com/jonaprieto/f1d29c826de66c728ec5ee1e1649ed84)
(however, it's pretty ad-hoc and needs double-checking as it changes not
just files but file/folder references in the file contents).
- trim trailing whitespaces
- no large files
- python scripts formatting and imports

Pending is to update CI with these checks; at the moment, you can run
the pre-commit checks with

```
pre-commit run --all-files
```

This PR is motivated by the ongoing integration with Isabelle/HOL. For
example, the current folder and file names make it impossible for this
system to import the generated files by Juvix.