Add RMv3 content (#131)
This is basically https://github.com/anoma/nspec/pull/109 "manually
rebased" on the `v2` branch. The only change is that I omitted the
proving system definition as we're already moved that elsewhere. This
content will need to be edited and cleaned up a bit in the future to fit
the rest of the specs but I wanted to get it in for now so that we can
clarify things for engineering.

---------

Co-authored-by: Anoma Research <arts@heliax.dev>
Co-authored-by: Yulia Khalniyazova <juliajarskilli@gmail.com>
Co-authored-by: Michael <20623991+heueristik@users.noreply.github.com>