Prose improvements for commitment, decryption, and identity management engines (#307)
Replace overview and purpose sections of the main engine files for
commitment, decryption, and identity management engines with something
that's, hopefully, more intuitive. I also redid all the flow charts to
be more detailed and added English descriptions of what each flowchart
depicts. It may make sense to move these charts to the top of the file.

---------

Co-authored-by: Jonathan Cubides <jonathan.cubides@uib.no>