Prose improvements for mempool worker, executor, and shard engines (#311)
Replace overview sections of the main engine files for mempool worker,
executor, and shard engines with something that's, hopefully, more
intuitive. Shards already had a detailed description, but I prepended it
with an intuitive overview that's more explicit about the message
structure. I did something similar with the mempool engine, though
there's a lot more overlap between the content of what I added and what
was already there. I'm hesitant to try just fusing them because of the
vigorous cross-linking that exists with the old text that I'm not even
sure works.

I also redid all the flow charts to be more detailed, and added english
descriptions of what each flowcharts depict. I also moved the charts to
the top of the file.

---------

Co-authored-by: Jonathan Cubides <jonathan.cubides@uib.no>