Prose improvements for encryption, verification, and naming engines (#335)
Replace overview sections of the main engine files for encryption,
verification, and naming engines with something that's, hopefully, more
intuitive. I also redid all the flow charts to be more detailed, and
added english descriptions of what each flowcharts depict. I also moved
the charts to the top of the file.

---------

Co-authored-by: Tobias Heindel <heindel@protonmail.com>
Co-authored-by: Jonathan Cubides <jonathan.cubides@uib.no>